api/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
sqlalchemy[asyncio]>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
pydantic[email]>=2.0.0
pydantic-settings>=2.6.0
python-jose[cryptography]>=3.3.0
httpx>=0.28.0
scipy>=1.14.0
numpy>=2.1.0
python-multipart>=0.0.12
